using Redis for faster chat message, adding, getting and updating

+++ services/chat/deleteChat.js
@@ -2,6 +2,7 @@ import deleteChatByIdDB from "../../database/Chat/deleteChatByIdDB.js";
 import catchGraphQLError from "../../lib/catchGraphQLError.js";
 import Req from "../../lib/Req.js";
 import socketConnect from "../../lib/socketConnect.js";
+import { deleteChatMsgIntoRedis } from "../../redis/Chat/chat.js";
 
 const deleteChat = catchGraphQLError(async (parent, args, { req, loaders }) => {
   const findUser = await Req(req);
@@ -10,7 +11,11 @@ const deleteChat = catchGraphQLError(async (parent, args, { req, loaders }) => {
 
   const { chatId, roomId } = args;
 
-  await deleteChatByIdDB(chatId);
+  const isDeletedFromRedis = await deleteChatMsgIntoRedis(chatId);
+
+  if (!isDeletedFromRedis) {
+    await deleteChatByIdDB(chatId);
+  }
 
   io.to(roomId).emit("delete-chat", { roomId, chatId });
